migrations to change price field to decimal

+++ api/models.py
@@ -5,8 +5,8 @@ from django.utils import timezone
 
 class PriceInMinutes(models.Model):
   timestamp = models.DateTimeField(auto_now_add=True, null=True, blank=True)
-  market_price = models.IntegerField(blank=False)
+  market_price = models.DecimalField(blank=False, max_digits=15, decimal_places=2)
 
 class HistoricalData(models.Model):
-  market_price = models.IntegerField(blank=False)
+  market_price = models.DecimalField(blank=False, max_digits=15, decimal_places=2)
   date = models.DateField(auto_now_add=False)
